Textures added to folders Added truck, RND amd farmer sounds (13) dog added added Resources (font and UISkin) baseActor Added added All Screens and navegation added Mock Added dialogs added spinActor and fixxed labels Full funcionality, no persistence, no DB full decoration

+++ core/src/com/mygdx/game/images/ImageClampToEdge.java
@@ -9,7 +9,7 @@ import com.mygdx.game.Constants;
 
 public class ImageClampToEdge extends Image {
 
-    public ImageClampToEdge(Texture texture, int x, int y,float width,float height){
+    public ImageClampToEdge(Texture texture, float x, float y,float width,float height){
         texture.setWrap(Texture.TextureWrap.ClampToEdge, Texture.TextureWrap.ClampToEdge);
         TextureRegion imgTextureRegion = new TextureRegion(texture);
         imgTextureRegion.setRegion(0,0,texture.getWidth(),texture.getHeight());
